refactor(FollowDto): Follow Dto 네이밍 컨벤션 정리
