+++ main-server/src/main/java/com/example/demo/domain/follow/dto/response/IsUserFollowingResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.domain.follow.dto;
+package com.example.demo.domain.follow.dto.response;
 
 import lombok.Builder;
